Rename types module to gprofiler_types to avoid imports errors
