+++ gprofiler/profilers/profiler_base.py
@@ -10,8 +10,8 @@ from typing import List, Optional
 from psutil import NoSuchProcess, Process
 
 from gprofiler.exceptions import StopEventSetException
+from gprofiler.gprofiler_types import ProcessToStackSampleCounters, StackToSampleCount
 from gprofiler.log import get_logger_adapter
-from gprofiler.types import ProcessToStackSampleCounters, StackToSampleCount
 from gprofiler.utils import limit_frequency
 
 logger = get_logger_adapter(__name__)
