+++ gprofiler/profilers/python.py
@@ -13,11 +13,11 @@ from typing import List, Optional
 from psutil import Process
 
 from gprofiler.exceptions import CalledProcessError, ProcessStoppedException, StopEventSetException
+from gprofiler.gprofiler_types import ProcessToStackSampleCounters, StackToSampleCount
 from gprofiler.log import get_logger_adapter
 from gprofiler.merge import parse_and_remove_one_collapsed, parse_many_collapsed
 from gprofiler.profilers.profiler_base import ProcessProfilerBase, ProfilerBase, ProfilerInterface
 from gprofiler.profilers.registry import register_profiler
-from gprofiler.types import ProcessToStackSampleCounters, StackToSampleCount
 from gprofiler.utils import (
     pgrep_maps,
     poll_process,
