+++ gprofiler/profilers/ruby.py
@@ -9,11 +9,11 @@ from typing import List
 from psutil import Process
 
 from gprofiler.exceptions import ProcessStoppedException, StopEventSetException
+from gprofiler.gprofiler_types import StackToSampleCount
 from gprofiler.log import get_logger_adapter
 from gprofiler.merge import parse_and_remove_one_collapsed
 from gprofiler.profilers.profiler_base import ProcessProfilerBase
 from gprofiler.profilers.registry import register_profiler
-from gprofiler.types import StackToSampleCount
 from gprofiler.utils import pgrep_maps, random_prefix, resource_path, run_process
 
 logger = get_logger_adapter(__name__)
